Extract notifier config, introduce alertmanager client config
